add likes/dislikes and delete functions

+++ src/utils/api.js
@@ -66,22 +66,40 @@ class Api {
       .catch(this._errorHandler)
   }
 
-  likeCard(cardId) {
-    return fetch(`${this.url}cards/likes/${cardId}`, {
-      method: 'PUT',
-      headers: this.headers
-    })
-      .then(this._resultHandler)
-      .catch(this._errorHandler)
-  }
+  // likeCard(cardId) {
+  //   return fetch(`${this.url}cards/likes/${cardId}`, {
+  //     method: 'PUT',
+  //     headers: this.headers
+  //   })
+  //     .then(this._resultHandler)
+  //     .catch(this._errorHandler)
+  // }
+  //
+  // dislikeCard(cardId) {
+  //   return fetch(`${this.url}cards/likes/${cardId}`, {
+  //     method: 'DELETE',
+  //     headers: this.headers
+  //     })
+  //     .then(this._resultHandler)
+  //     .catch(this._errorHandler)
+  // }
 
-  dislikeCard(cardId) {
-    return fetch(`${this.url}cards/likes/${cardId}`, {
-      method: 'DELETE',
-      headers: this.headers
-      })
-      .then(this._resultHandler)
-      .catch(this._errorHandler)
+  changeLikeCardStatus (cardId, isLiked) {
+    if (isLiked) {
+      return fetch(`${this.url}cards/likes/${cardId}`, {
+            method: 'PUT',
+            headers: this.headers
+          })
+            .then(this._resultHandler)
+            .catch(this._errorHandler)
+        } else {
+        return fetch(`${this.url}cards/likes/${cardId}`, {
+          method: 'DELETE',
+          headers: this.headers
+          })
+          .then(this._resultHandler)
+          .catch(this._errorHandler)
+    }
   }
 
   deleteCard(cardId) {
